remove road_class model and finish categorical model
