+++ rust/routee-compass-core/src/model/constraint/default/categorical/categorical_model.rs
@@ -1,19 +1,40 @@
-use crate::model::constraint::ConstraintModel;
-struct CategoricalConstraintModel {}
+use super::categorical_service::CategoricalModelService;
+use crate::model::constraint::{ConstraintModel, ConstraintModelError};
+use crate::model::network::Edge;
+use crate::model::state::StateModel;
+use crate::model::state::StateVariable;
+use crate::model::traversal::EdgeFrontierContext;
+use std::collections::HashSet;
+use std::sync::Arc;
+pub struct CategoricalConstraintModel {
+    pub service: Arc<CategoricalModelService>,
+    pub query_categories: Option<HashSet<u8>>,
+}
 
 impl ConstraintModel for CategoricalConstraintModel {
     fn valid_frontier(
         &self,
-        ctx: &crate::model::traversal::EdgeFrontierContext,
-        state: &[crate::model::state::StateVariable],
-        state_model: &crate::model::state::StateModel,
-    ) -> Result<bool, crate::model::constraint::ConstraintModelError> {
-        todo!();
+        ctx: &EdgeFrontierContext,
+        _state: &[StateVariable],
+        _state_model: &StateModel,
+    ) -> Result<bool, ConstraintModelError> {
+        self.valid_edge(ctx.edge)
     }
-    fn valid_edge(
-        &self,
-        edge: &crate::model::network::Edge,
-    ) -> Result<bool, crate::model::constraint::ConstraintModelError> {
-        todo!();
+
+    fn valid_edge(&self, edge: &Edge) -> Result<bool, ConstraintModelError> {
+        match &self.query_categories {
+            None => Ok(true),
+            Some(encoding) => self
+                .service
+                .category_by_edge
+                .get(edge.edge_id.0)
+                .ok_or_else(|| {
+                    ConstraintModelError::ConstraintModelError(format!(
+                        "edge id {} missing from constraint model file",
+                        edge.edge_id
+                    ))
+                })
+                .map(|id| encoding.contains(id)),
+        }
     }
 }
