+++ rust/routee-compass-core/src/model/constraint/default/categorical/categorical_service.rs
@@ -1,11 +1,18 @@
+use super::categorical_model::CategoricalConstraintModel;
 use crate::model::{
     constraint::{ConstraintModel, ConstraintModelError, ConstraintModelService},
     state::StateModel,
 };
 use serde_json;
-use std::{collections::HashMap, sync::Arc};
+use serde_json::Value;
+use std::{
+    collections::{HashMap, HashSet},
+    sync::Arc,
+};
+
+#[derive(Clone)]
 pub struct CategoricalModelService {
-    pub key: Arc<String>,                           // the category name
+    pub key: String,                                // the category name
     pub category_by_edge: Arc<Box<[u8]>>,           // categorizes each edge with the encoding
     pub category_mapping: Arc<HashMap<String, u8>>, // maps the categories (String) to the encoding (u8)
 }
@@ -14,8 +21,61 @@ impl ConstraintModelService for CategoricalModelService {
     fn build(
         &self,
         query: &serde_json::Value,
-        state_model: Arc<StateModel>,
+        _state_model: Arc<StateModel>,
     ) -> Result<Arc<dyn ConstraintModel>, ConstraintModelError> {
-        todo!();
+        let query_categories = match query
+            .get(&self.key)
+            .map(|val| read_categories_from_query(val, &self.key))
+        {
+            Some(Err(e)) => Err(e),
+            Some(Ok(categories)) => {
+                let mapped: Result<HashSet<u8>, ConstraintModelError> = categories
+                    .iter()
+                    .map(|c| {
+                        self.category_mapping.get(c).copied().ok_or_else(|| {
+                            ConstraintModelError::BuildError(format!(
+                                "road class '{}' not found in road class mapping",
+                                c
+                            ))
+                        })
+                    })
+                    .collect();
+                mapped.map(Some)
+            }
+            None => Ok(None),
+        }?;
+
+        let service: Arc<CategoricalModelService> = Arc::new(self.clone());
+        let model = CategoricalConstraintModel {
+            service,
+            query_categories,
+        };
+        Ok(Arc::new(model))
     }
+}
+
+fn read_categories_from_query(
+    value: &Value,
+    category_key: &String,
+) -> Result<HashSet<String>, ConstraintModelError> {
+    let arr = value.as_array().ok_or_else(|| {
+        ConstraintModelError::BuildError(format!(
+            "query's {category_key} value must be an array, found '{value}'"
+        ))
+    })?;
+    // if the value is a string (or number or bool), store it as a valid road class
+    let arr_str = arr
+        .iter()
+        .enumerate()
+        .map(|(idx, c)| match c {
+            Value::Bool(b) => Ok(b.to_string()),
+            Value::Number(number) => Ok(number.to_string()),
+            Value::String(string) => Ok(string.clone()),
+            _ => Err(ConstraintModelError::BuildError(format!(
+                "query's '{category_key}[{idx}]' value must be a string, found '{c}'"
+            ))),
+        })
+        .collect::<Result<HashSet<_>, _>>()?;
+
+    Ok(arr_str)
 }
