+++ rust/routee-compass-core/src/model/constraint/default/categorical/mod.rs
@@ -1,4 +1,4 @@
-mod categorical_builder;
+pub mod categorical_builder;
 mod categorical_builder_config;
 mod categorical_model;
 mod categorical_service;
