+++ rust/routee-compass-core/src/model/constraint/default/mod.rs
@@ -2,6 +2,5 @@ pub mod categorical;
 pub mod combined;
 pub mod no_restriction;
 pub mod no_restriction_builder;
-pub mod road_class;
 pub mod turn_restrictions;
 pub mod vehicle_restrictions;
