--- rust/routee-compass-core/src/model/constraint/default/road_class/mod.rs
@@ -1,4 +1,0 @@
-pub mod road_class_builder;
-pub mod road_class_builder_config;
-pub mod road_class_model;
-pub mod road_class_service;
