--- rust/routee-compass-core/src/model/constraint/default/road_class/road_class_builder_config.rs
@@ -1,7 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(deny_unknown_fields)]
-pub struct RoadClassBuilderConfig {
-    pub road_class_input_file: String,
-}
